requirements.txt
Flask
python-dotenv
tapo
requests
re